refacto the code structure and types
