+++ sandbox/src/app/types/OtelConfig.type.ts
@@ -1,0 +1,6 @@
+export interface OtelConfig {
+  serviceName: string;
+  serviceVersion: string;
+  tracesUrl: string;
+  logsUrl: string;
+}
